refactor: group path and file manip in pathutils
As the code is becoming increasingly long and difficult to understand,
I feel it is a good idea to start refactoring it a bit and split the
various utilities in separate functions and dirs.

This commit introduces `pathutils`, whose purpose is to regroup all
functions related to path and file manipulation (sanitizing, removal of
extension, etc.)
